Better arrows in options UI

+++ soh/include/z64.h
@@ -1502,7 +1502,7 @@ typedef struct {
     uint8_t bossRushIndex;
     uint8_t bossRushOffset;
     int16_t bossRushUIAlpha;
-    float bossRushArrowOffset;
+    uint16_t bossRushArrowOffset;
 } FileChooseContext; // size = 0x1CAE0
 
 typedef enum {
